NXP-28758: factor common Session code into BaseSession
